requirements.txt
dask==0.13.0
decorator==4.0.10
fonttools==3.4.0
funcsigs==1.0.2
h5py==2.6.0
Keras==1.2.0
mock==2.0.0
networkx==1.11
numpy==1.11.3
olefile==0.43
pbr==1.10.0
Pillow==4.0.0
protobuf==3.1.0.post1
PyYAML==3.12
scikit-image==0.12.3
scipy==0.18.1
six==1.10.0
tensorflow==0.12.1
Theano==0.8.2
toolz==0.8.2

